Conflicts:
	src/libcore/fmt/mod.rs
	src/libcore/hash/mod.rs
	src/libcore/iter.rs
	src/libcore/kinds.rs
	src/librustc/middle/expr_use_visitor.rs
	src/librustc/middle/infer/combine.rs
	src/librustc_typeck/check/_match.rs
	src/librustc_typeck/check/demand.rs
	src/librustc_typeck/variance.rs
	src/libstd/io/mod.rs
	src/libstd/sync/atomic.rs
	src/libsyntax/ext/format.rs
	src/libsyntax/std_inject.rs
	src/test/auxiliary/default_type_params_xc.rs
	src/test/auxiliary/issue-2380.rs
	src/test/auxiliary/lang-item-public.rs
	src/test/auxiliary/nested_item.rs
	src/test/compile-fail/destructure-trait-ref.rs
	src/test/compile-fail/dst-object-from-unsized-type.rs
	src/test/compile-fail/generic-impl-less-params-with-defaults.rs
	src/test/compile-fail/generic-impl-more-params-with-defaults.rs
	src/test/compile-fail/generic-type-less-params-with-defaults.rs
	src/test/compile-fail/generic-type-more-params-with-defaults.rs
	src/test/compile-fail/generic-type-params-name-repr.rs
	src/test/compile-fail/issue-13058.rs
	src/test/compile-fail/issue-17551.rs
	src/test/compile-fail/issue-17636.rs
	src/test/compile-fail/issue-6596-2.rs
	src/test/compile-fail/kindck-impl-type-params.rs
	src/test/compile-fail/lifetime-inference-give-expl-lifetime-param-2.rs
	src/test/compile-fail/lifetime-inference-give-expl-lifetime-param-3.rs
	src/test/compile-fail/lifetime-inference-give-expl-lifetime-param.rs
	src/test/compile-fail/liveness-use-after-send.rs
	src/test/compile-fail/privacy4.rs
	src/test/compile-fail/regions-close-object-into-object.rs
	src/test/compile-fail/required-lang-item.rs
	src/test/compile-fail/typeck_type_placeholder_mismatch.rs
	src/test/compile-fail/unboxed-closure-sugar-default.rs
	src/test/compile-fail/unboxed-closure-sugar-equiv.rs
	src/test/compile-fail/unboxed-closure-sugar-lifetime-elision.rs
	src/test/compile-fail/unboxed-closure-sugar-region.rs
	src/test/compile-fail/unsized3.rs
	src/test/compile-fail/unsized6.rs
	src/test/compile-fail/unsized7.rs
	src/test/compile-fail/variance-regions-direct.rs
	src/test/compile-fail/variance-regions-indirect.rs
	src/test/compile-fail/variance-regions-unused-direct.rs
	src/test/run-pass/associated-types-qualified-path-with-trait-with-type-parameters.rs
	src/test/run-pass/closure-syntax.rs
	src/test/run-pass/issue-14589.rs
	src/test/run-pass/issue-17636.rs
	src/test/run-pass/issue-6157.rs
	src/test/run-pass/method-recursive-blanket-impl.rs
	src/test/run-pass/trait-object-generics.rs
	src/test/run-pass/unsized.rs
	src/test/run-pass/unsized2.rs

+++ src/test/compile-fail/borrowck-init-in-fru.rs
@@ -8,10 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#[deriving(Clone)]
+#[derive(Clone)]
 struct point {
-    x: int,
-    y: int,
+    x: isize,
+    y: isize,
 }
 
 fn main() {
