+++ src/test/compile-fail/borrowck-move-out-of-static-item.rs
@@ -10,10 +10,10 @@
 
 // Ensure that moves out of static items is forbidden
 
-use std::kinds::marker;
+use std::marker;
 
 struct Foo {
-    foo: int,
+    foo: isize,
     nocopy: marker::NoCopy
 }
