+++ src/test/compile-fail/issue-13359.rs
@@ -13,9 +13,9 @@ fn foo(_s: i16) { }
 fn bar(_s: u32) { }
 
 fn main() {
-    foo(1*(1 as int));
-    //~^ ERROR: mismatched types: expected `i16`, found `int` (expected i16, found int)
+    foo(1*(1 as isize));
+    //~^ ERROR: mismatched types: expected `i16`, found `isize` (expected i16, found isize)
 
-    bar(1*(1 as uint));
-    //~^ ERROR: mismatched types: expected `u32`, found `uint` (expected u32, found uint)
+    bar(1*(1 as usize));
+    //~^ ERROR: mismatched types: expected `u32`, found `usize` (expected u32, found usize)
 }
