+++ src/test/compile-fail/issue-18611.rs
@@ -8,12 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![feature(associated_types)]
-
-fn add_state(op:
-    <int as HasState>::State
-//~^ ERROR it is currently unsupported to access associated types except through a type parameter
-) {}
+fn add_state(op: <isize as HasState>::State) {
+//~^ ERROR the trait `HasState` is not implemented for the type `isize`
+}
 
 trait HasState {
     type State;
