+++ src/test/compile-fail/issue-4366-2.rs
@@ -8,8 +8,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![feature(globs)]
-
 // ensures that 'use foo:*' doesn't import non-public item
 
 use m1::*;
@@ -20,7 +18,7 @@ mod foo {
 mod a {
     pub mod b {
         use foo::foo;
-        type bar = int;
+        type bar = isize;
     }
     pub mod sub {
         use a::b::*;
@@ -36,4 +34,3 @@ mod m1 {
 fn main() {
     foo(); //~ ERROR: unresolved name
 }
-
