+++ src/test/compile-fail/issue-4366.rs
@@ -13,8 +13,6 @@
 // ensures that 'use foo:*' doesn't import non-public 'use' statements in the
 // module 'foo'
 
-#![feature(globs)]
-
 use m1::*;
 
 mod foo {
@@ -23,11 +21,11 @@ mod foo {
 mod a {
     pub mod b {
         use foo::foo;
-        type bar = int;
+        type bar = isize;
     }
     pub mod sub {
         use a::b::*;
-        fn sub() -> int { foo(); 1 } //~ ERROR: unresolved name `foo`
+        fn sub() -> isize { foo(); 1 } //~ ERROR: unresolved name `foo`
     }
 }
