+++ src/test/compile-fail/map-types.rs
@@ -8,6 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(box_syntax)]
+
 extern crate collections;
 
 use std::collections::HashMap;
@@ -22,8 +24,8 @@ impl<K, V> Map<K, V> for HashMap<K, V> {}
 // Test that trait types printed in error msgs include the type arguments.
 
 fn main() {
-    let x: Box<HashMap<int, int>> = box HashMap::new();
-    let x: Box<Map<int, int>> = x;
-    let y: Box<Map<uint, int>> = box x;
-    //~^ ERROR the trait `Map<uint, int>` is not implemented
+    let x: Box<HashMap<isize, isize>> = box HashMap::new();
+    let x: Box<Map<isize, isize>> = x;
+    let y: Box<Map<usize, isize>> = box x;
+    //~^ ERROR the trait `Map<usize, isize>` is not implemented
 }
