+++ src/test/compile-fail/nonscalar-cast.rs
@@ -10,11 +10,11 @@
 
 // error-pattern:non-scalar cast
 
-#[deriving(Show)]
+#[derive(Show)]
 struct foo {
-    x: int
+    x: isize
 }
 
 fn main() {
-    println!("{}", foo{ x: 1 } as int);
+    println!("{}", foo{ x: 1 } as isize);
 }
