+++ src/test/compile-fail/static-items-cant-move.rs
@@ -10,10 +10,10 @@
 
 // Verifies that static items can't be moved
 
-use std::kinds::marker;
+use std::marker;
 
 struct Foo {
-    foo: int,
+    foo: isize,
     nocopy: marker::NoCopy
 }
